+++ gssh/pkg/aws/ip.go
@@ -1,4 +1,4 @@
-package gssh
+package aws
 
 import (
 	"fmt"
