+++ gssh/pkg/aws/ip_test.go
@@ -1,4 +1,4 @@
-package gssh
+package aws
 
 import (
 	"testing"
