+++ gssh/pkg/ssh/ssh_test.go
@@ -1,11 +1,11 @@
-package gssh
+package ssh
 
 import (
 	"testing"
 )
 
 func TestGetPublicKey(t *testing.T) {
-	file := "../../test/id_rsa"
+	file := "../../testdata/id_rsa"
 	_, err := GetPublicKey(file)
 	if err != nil {
 		t.Error("Not able to read public key file")
