+++ gssh/testdata/id_rsa.pub
